// src/modules/jumpseller/endpoints.enum.ts
export enum JumpsellerEndpoints {
  PRODUCTS = "products.json",
  PRODUCT = "products",
  CUSTOM_FIELDS = "custom_fields.json",
  UPDATE_CUSTOM_FIELDS = "custom_fields",
  FIELDS = "fields.json",
  STOCK = "products_locations",
  VARIANTS = "variants.json",
  VARIANT = "variants",
  IMAGES = "images.json",
}